Factor getTables out of QueryBuilderRoute
